new controller added
