+++ src/test/java/com/springreact/Spring_React/HelloControllerTest.java
@@ -1,10 +1,10 @@
 package com.springreact.Spring_React;
 
+import com.springreact.Spring_React.controller.HelloController;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
 import java.nio.charset.StandardCharsets;
